+++ src/output/p_mp3.h
@@ -18,6 +18,7 @@
 #include "common/byte_buffer.h"
 #include "common/mp3.h"
 #include "merge/generic_packetizer.h"
+#include "merge/stream_property_preserver.h"
 #include "merge/timestamp_calculator.h"
 
 class mp3_packetizer_c: public generic_packetizer_c {
@@ -28,6 +29,7 @@ private:
   mtx::bytes::buffer_c m_byte_buffer;
   bool m_codec_id_set, m_valid_headers_found;
   timestamp_calculator_c m_timestamp_calculator;
+  stream_property_preserver_c<timestamp_c> m_discard_padding;
   int64_t m_packet_duration;
   std::vector<packet_extension_cptr> m_packet_extensions;