+++ tests/test-644mp3_with_discard_padding.rb
@@ -1,0 +1,5 @@
+#!/usr/bin/ruby -w
+
+# T_644mp3_with_discard_padding
+describe "mkvmerge / MP3 with discard padding"
+test_merge "data/mkv/mp3-with-discard-padding.mka"